+++ src/FlashBeat.ExternalServices/Dtos/Responses/TrackBaseDto.cs
@@ -8,7 +8,7 @@ public record TrackBaseDto
     /// <summary>
     /// Gets the id of the track.
     /// </summary>
-    public required int Id { get; init; }
+    public required long Id { get; init; }
 
     /// <summary>
     /// Gets the title of the track.
