+++ src/FlashBeat.Web.Service/Services/Tracks/Responses/ArtistView.cs
@@ -12,7 +12,7 @@ public sealed record ArtistView
     /// <summary>
     /// Gets the id of the artist.
     /// </summary>
-    public required int Id { get; init; }
+    public required long Id { get; init; }
 
     /// <summary>
     /// Gets the name of the artist.
